+++ src/redis_bitmap_string.h
@@ -14,10 +14,10 @@ class BitmapString : public Database {
   rocksdb::Status GetBit(const std::string &raw_value, uint32_t offset, bool *bit);
   rocksdb::Status SetBit(const Slice &ns_key, std::string *raw_value, uint32_t offset, bool new_bit, bool *old_bit);
   rocksdb::Status BitCount(const std::string &raw_value, int start, int stop, uint32_t *cnt);
-  rocksdb::Status BitPos(const std::string &raw_value, bool bit, int start, int stop, bool stop_given, int *pos);
+  rocksdb::Status BitPos(const std::string &raw_value, bool bit, int start, int stop, bool stop_given, int64_t *pos);
  private:
   size_t redisPopcount(unsigned char *p, int count);
-  int redisBitpos(unsigned char *c, int count, int bit);
+  int64_t redisBitpos(unsigned char *c, int count, int bit);
 };
 
 }  // namespace Redis
